Moved the new TPP for gn fwd and bwd into a separate header, removed most of the macro there
