+++ version.txt
@@ -1,1 +1,1 @@
-dev/kvoronin/groupnormalization_tpp-1.17-2142
+dev/kvoronin/groupnormalization_tpp-1.17-2143
